+++ ros/drake_ros_systems/include/drake/ros/systems/ros_sensor_publisher_odometry.h
@@ -5,11 +5,11 @@
 #include "ros/ros.h"
 #include "nav_msgs/Odometry.h"
 
-#include "drake/systems/System.h"
+#include "drake/system1/System.h"
 #include "drake/systems/plants/KinematicsCache.h"
 #include "drake/systems/plants/RigidBodyTree.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/systems/vector.h"
+#include "drake/system1/vector.h"
 
 using drake::NullVector;
 using drake::RigidBodySensor;